requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sse-starlette==1.6.5
aiofiles==23.2.1
python-multipart==0.0.6
pydantic==2.7.2
requests==2.31.0
pywin32==306; sys_platform == 'win32'
pypiwin32==223; sys_platform == 'win32'
jinja2==3.1.3
httpx==0.27.0
python-dotenv==1.0.0
fastapi-mcp==0.3.0 